--- Weitianshi/pages/discover/discoverPerson/diccoverPerson.json
@@ -1,1 +1,0 @@
-{}
