--- Weitianshi/pages/discover/discoverPerson/diccoverPerson.wxss
@@ -1,1 +1,0 @@
-/* pages/discover/discoverPerson/diccoverPerson.wxss */
