+++ Weitianshi/pages/discover/discoverPerson/discoverPerson.json
@@ -1,0 +1,5 @@
+{
+    "navigationBarTitleText": "投资人",
+    "navigationBarBackgroundColor": "#fafafa",
+    "navigationBarTextStyle": "black"
+  }
