+++ Weitianshi/pages/discover/discoverPerson/discoverPerson.wxss
@@ -1,0 +1,379 @@
+/*tab切换页  */
+
+.swiper-tab {
+  /* width: 88%; */
+  line-height: 80rpx;
+  background: #40587a;
+  z-index: 999;
+  padding-left: 32rpx;
+}
+
+.swiper-tab-list {
+  display: inline-block;
+  /* padding-left: 32rpx; */
+  margin-right: 58rpx;
+  font-size: 30rpx;
+  color: #999;
+  line-height: 62rpx;
+  text-align: center;
+  border: none;
+}
+
+.swiper-tab-list:last-child {
+  margin-right: 0;
+}
+
+.swiper-box {
+  height: 945rpx;
+}
+
+.swiper-box-other {
+  height: 1034rpx;
+}
+
+.on {
+  color: white;
+  text-align: center;
+  font-size: 30rpx;
+  line-height: 80rpx;
+  border-bottom: 4rpx solid white;
+}
+
+.dropDown {
+  top: 160rpx;
+}
+
+/*banner轮播图  */
+
+.slide-image {
+  width: 100%;
+  height: 100%;
+}
+
+/*精选项目copy  */
+
+.browseMeTitle {
+  /*height: 90rpx;*/
+  margin: 32rpx 32rpx 32rpx 32rpx;
+}
+
+.browseMeTitle .num {
+  font-size: 24rpx;
+  color: #666;
+  line-height: 24rpx;
+  text-align: right;
+}
+
+.title {
+  font-size: 26rpx;
+  color: #333;
+  line-height: 26rpx;
+  text-align: left;
+}
+
+.title .block {
+  background: #40587a;
+  width: 4rpx;
+  height: 24rpx;
+  margin-right: 12rpx;
+}
+
+.contactsPic {
+  width: 750rpx;
+  height: 280rpx;
+}
+
+/*认证细腻  */
+
+.accreditation {
+  background: #fff1e8;
+  height: 88rpx;
+  font-size: 24rpx;
+  color: #999;
+  line-height: 88rpx;
+  text-align: left;
+  padding-left: 32rpx;
+}
+
+.toAccreditation {
+  font-size: 24rpx;
+  color: #fff;
+  line-height: 52rpx;
+  text-align: center;
+  background: #40587a;
+  border-radius: 4rpx;
+  width: 136rpx;
+  height: 52rpx;
+  margin-top: 18rpx;
+  position: absolute;
+  right: 32rpx;
+}
+
+.push {
+  height: 88rpx;
+  font-size: 28rpx;
+  line-height: 88rpx;
+  background: #40587a;
+  text-align: center;
+  color: #fff;
+}
+
+.pushing {
+  background: #ccc;
+  height: 88rpx;
+  font-size: 28rpx;
+  line-height: 88rpx;
+  text-align: center;
+  color: #fff;
+}
+
+.aBox {
+  height: 32rpx;
+  width: 100%;
+}
+
+.t-highBox {
+  height: 88rpx;
+  width: 100%;
+}
+
+.searchNothing {
+  font-size: 24rpx;
+  color: #999;
+  letter-spacing: 0;
+  line-height: 24rpx;
+  text-align: center;
+  margin-top: 252rpx;
+}
+
+.searchLine {
+  margin-bottom: 72rpx;
+  height: 2rpx;
+  margin-left: 32rpx;
+  margin-right: 32rpx;
+  border-bottom: 2rpx dashed #999;
+  width: 90%;
+}
+
+.cardBottom {
+  margin: 0;
+  font-size: 28rpx;
+  color: #40587a;
+  margin-bottom: 24rpx;
+}
+
+.bigTitle {
+  height: 280rpx;
+  width: 100%;
+  position: relative;
+}
+
+.searchImageBox {
+  background: #009eff;
+  width: 120rpx;
+  height: 76rpx;
+  position: absolute;
+  top: 0;
+  right: 0;
+  border-radius: 4rpx;
+}
+
+.searchImage {
+  height: 40rpx;
+  width: 40rpx;
+  position: absolute;
+  top: 19rpx;
+  right: 39rpx;
+}
+
+.swiper-box .bigTitle .searchSth {
+  text-align: left;
+}
+
+.searchSth {
+  background: #fff;
+  border-radius: 4rpx;
+  width: 596rpx;
+  height: 18rpx;
+  position: absolute;
+  top: 162rpx;
+  left: 62rpx;
+  font-size: 24rpx;
+  color: #999;
+  line-height: 24rpx;
+  text-align: left;
+  padding: 28rpx 0 30rpx 26rpx;
+}
+
+.outStyleBox {
+  margin-bottom: 32rpx;
+  background: #fff;
+  padding: 32rpx;
+}
+.outStyleBox1 {
+  margin-bottom: 32rpx;
+  background: #fff;
+}
+.outPadding{
+  padding: 32rpx 32rpx 0 32rpx;
+}
+.outStyleBox:last-child {
+  margin-bottom: 0;
+}
+
+/* 投资风格 */
+
+.blockTitle {
+  font-size: 30rpx;
+  color: #333;
+  line-height: 33rpx;
+  text-align: left;
+  margin-bottom: 16rpx;
+}
+
+.blockIntro {
+  font-size: 20rpx;
+  color: #999;
+  line-height: 22rpx;
+  text-align: left;
+  margin-bottom: 40rpx;
+}
+
+.investmentListBox {
+  flex-wrap: wrap;
+}
+
+.investStyleBlock {
+  background: #fff;
+  box-shadow: 0 0 12rpx 0 rgba(64, 88, 122, 0.10);
+  border-radius: 8rpx;
+  width: 220rpx;
+  height: 152rpx;
+  margin-bottom: 16rpx;
+  margin-right: 9rpx;
+}
+
+.investStyleTitle {
+  font-size: 26rpx;
+  color: #333;
+  line-height: 26rpx;
+  padding-top: 44rpx;
+  padding-bottom: 16rpx;
+}
+
+.investStyleNum {
+  font-size: 22rpx;
+  color: #40587a;
+  line-height: 25rpx;
+  font-weight: bold;
+}
+
+.swiper-box-other .outStyleBox view {
+  text-align: left;
+}
+
+.swiper-box-other .outStyleBox .investStyleBlock view {
+  text-align: center;
+}
+
+.investStyleWord {
+  color: #999;
+  font-weight: lighter;
+  font-size: 22rpx;
+  line-height: 25rpx;
+}
+
+/*热门领域  */
+
+.industryTitleBlock {
+  font-size: 22rpx;
+  color: #999;
+  line-height: 22rpx;
+  text-align: left;
+}
+
+.industryTitle {
+  margin-right: 40rpx;
+  margin-bottom: 32rpx;
+  white-space: nowrap;
+  color: #666;
+}
+
+.industryTitle:first-child {
+  margin-right: 258rpx;
+}
+
+.industryTitle:last-child {
+  margin-right: 0;
+}
+
+.industryContentBlock {
+  font-size: 28rpx;
+  color: #333;
+  line-height: 80rpx;
+}
+
+.industryContentBlock:last-child {
+  margin-bottom: 0;
+}
+
+.industryContentTitle {
+  width: 304rpx;
+  margin-left: 36rpx;
+}
+
+.industryContent {
+  width: 128rpx;
+  color: #666;
+}
+
+.industryContent:last-child {
+  width: 120rpx;
+}
+
+.allThing {
+  font-size: 24rpx;
+  color: #009eff;
+  line-height: 26rpx;
+  text-align: right;
+}
+
+.title {
+  justify-content: space-between;
+}
+
+.down {
+  font-size: 20rpx;
+  color: #28dc41;
+  line-height: 70rpx;
+  text-align: left;
+}
+
+.up {
+  font-size: 20rpx;
+  color: #f44c4c;
+  line-height: 70rpx;
+  text-align: left;
+}
+
+.arrow {
+  width: 12rpx;
+  height: 6rpx;
+  padding-top: 12rpx;
+  margin: 18rpx 4rpx 0 12rpx;
+}
+
+.high {
+  height: 88rpx;
+  width: 100%;
+}
+
+.industryNum {
+  width: 80rpx;
+  margin-right: 5rpx;
+}
+.t-investorColor{
+  background: #fff;
+  /* margin: 0 30rpx 0 30rpx; */
+  margin-bottom: 32rpx
+}
