+++ Weitianshi/pages/message/contactProject/projectList/projectList.wxss
@@ -8,7 +8,15 @@
 }
 
 .t-noLight {
+  font-family: PingFangSC-Regular;
+  font-size: 24rpx;
   color: #666;
+  line-height: 26rpx;
+  text-align: left;
+  padding: 8rpx 16rpx;
+  background: #f5f5f5;
+  border-radius: 2px;
+  margin-right: 30rpx;
 }
 
 .browseMeTitle {
@@ -63,7 +71,7 @@
 }
 
 .hoverBack {
-   background: #E0E0E0; 
+  background: #e0e0e0;
 }
 
 #t-introduce {
@@ -289,7 +297,7 @@
 .btnPic {
   height: 28rpx;
   width: 28rpx;
-  margin-left:90rpx;
+  margin-left: 90rpx;
   margin-right: 16rpx;
   margin-top: 27rpx;
 }
@@ -355,12 +363,14 @@
   text-align: center;
   width: 50%;
 }
+
 .aline {
   margin: 0 32rpx;
   width: 90%;
   height: 2rpx;
   background: #e6e6e6;
 }
+
 .t-line {
   width: 2rpx;
   height: 35rpx;
